refactor(ErrorCode): 에러코드 파일명 변경
새로운 ErroCode 인터페이스를 선언하기 위해 파일명 변경
- ErrorCode를 ErrorCodeBefore로 임시로 이름 수정, 그에 따른 변경사항들

related to: #276
