+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/AlreadyExistsException.java
@@ -1,10 +1,10 @@
 package com.prograngers.backend.exception.badrequest;
 
 import com.prograngers.backend.exception.CustomException;
-import com.prograngers.backend.exception.ErrorCode;
+import com.prograngers.backend.exception.ErrorCodeBefore;
 
 public class AlreadyExistsException extends CustomException {
-    public AlreadyExistsException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
+    public AlreadyExistsException(ErrorCodeBefore errorCodeBefore, String message) {
+        super(errorCodeBefore, message);
     }
 }
