+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/DifferentCodeLineNumberException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.DIFFERENT_SOLUTION;
+import static com.prograngers.backend.exception.ErrorCodeBefore.DIFFERENT_SOLUTION;
 
 public class DifferentCodeLineNumberException extends InvalidValueException {
     public DifferentCodeLineNumberException() {
