+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/DifferentSolutionException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.DIFFERENT_SOLUTION;
+import static com.prograngers.backend.exception.ErrorCodeBefore.DIFFERENT_SOLUTION;
 
 public class DifferentSolutionException extends InvalidValueException {
     public DifferentSolutionException() {
