+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/InvalidNotificationTypeException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.INVALID_NOTIFICATION_TYPE;
+import static com.prograngers.backend.exception.ErrorCodeBefore.INVALID_NOTIFICATION_TYPE;
 
 public class InvalidNotificationTypeException extends InvalidValueException {
     public InvalidNotificationTypeException() {
