+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/InvalidPageNumberException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.INVALID_PAGE_NUMBER;
+import static com.prograngers.backend.exception.ErrorCodeBefore.INVALID_PAGE_NUMBER;
 
 public class InvalidPageNumberException extends InvalidValueException {
     public InvalidPageNumberException() {
