+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/InvalidPageSizeException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.PAGE_SIZE_UNDER_THAN_ONE;
+import static com.prograngers.backend.exception.ErrorCodeBefore.PAGE_SIZE_UNDER_THAN_ONE;
 
 public class InvalidPageSizeException extends InvalidValueException{
     public InvalidPageSizeException() {
