+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/InvalidParentException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.INVALID_PARENT;
+import static com.prograngers.backend.exception.ErrorCodeBefore.INVALID_PARENT;
 
 public class InvalidParentException extends InvalidValueException {
     public InvalidParentException() {
