+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/LikesAlreadyExistsException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.ALREADY_EXISTS_LIKE;
+import static com.prograngers.backend.exception.ErrorCodeBefore.ALREADY_EXISTS_LIKE;
 
 public class LikesAlreadyExistsException extends AlreadyExistsException {
