+++ backend/src/main/java/com/prograngers/backend/exception/badrequest/PrivateSolutionException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.badrequest;
 
-import static com.prograngers.backend.exception.ErrorCode.PRIVATE_SOLUTION;
+import static com.prograngers.backend.exception.ErrorCodeBefore.PRIVATE_SOLUTION;
 
 public class PrivateSolutionException extends InvalidValueException {
