+++ backend/src/main/java/com/prograngers/backend/exception/enumtype/AlgorithmNotFoundException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.enumtype;
 
-import static com.prograngers.backend.exception.ErrorCode.ALGORITHM_NOT_EXISTS;
+import static com.prograngers.backend.exception.ErrorCodeBefore.ALGORITHM_NOT_EXISTS;
 
 public class AlgorithmNotFoundException extends EnumTypeException {
