+++ backend/src/main/java/com/prograngers/backend/exception/enumtype/EnumTypeException.java
@@ -1,13 +1,13 @@
 package com.prograngers.backend.exception.enumtype;
 
 import com.prograngers.backend.exception.CustomException;
-import com.prograngers.backend.exception.ErrorCode;
+import com.prograngers.backend.exception.ErrorCodeBefore;
 import lombok.Getter;
 
 @Getter
 public class EnumTypeException extends CustomException {
 
-    public EnumTypeException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
+    public EnumTypeException(ErrorCodeBefore errorCodeBefore, String message) {
+        super(errorCodeBefore, message);
     }
 }
