+++ backend/src/main/java/com/prograngers/backend/exception/enumtype/LanguageNotFoundException.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.exception.enumtype;
 
 
-import static com.prograngers.backend.exception.ErrorCode.LANGUAGE_NOT_EXISTS;
+import static com.prograngers.backend.exception.ErrorCodeBefore.LANGUAGE_NOT_EXISTS;
 
 public class LanguageNotFoundException extends EnumTypeException {
