+++ backend/src/main/java/com/prograngers/backend/exception/enumtype/LevelNotFoundException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.enumtype;
 
-import static com.prograngers.backend.exception.ErrorCode.LEVEL_NOT_EXISTS;
+import static com.prograngers.backend.exception.ErrorCodeBefore.LEVEL_NOT_EXISTS;
 
 public class LevelNotFoundException extends EnumTypeException {
