+++ backend/src/main/java/com/prograngers/backend/exception/enumtype/SortTypeNotFoundException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.enumtype;
 
-import static com.prograngers.backend.exception.ErrorCode.SORT_TYPE_NOT_EXISTS;
+import static com.prograngers.backend.exception.ErrorCodeBefore.SORT_TYPE_NOT_EXISTS;
 
 public class SortTypeNotFoundException extends EnumTypeException {
