+++ backend/src/main/java/com/prograngers/backend/exception/notfound/CommentAlreadyDeletedException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.notfound;
 
-import static com.prograngers.backend.exception.ErrorCode.COMMENT_ALREADY_DELETED;
+import static com.prograngers.backend.exception.ErrorCodeBefore.COMMENT_ALREADY_DELETED;
 
 public class CommentAlreadyDeletedException extends NotFoundException {
     public CommentAlreadyDeletedException() {
