+++ backend/src/main/java/com/prograngers/backend/exception/notfound/CommentNotFoundException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.notfound;
 
-import static com.prograngers.backend.exception.ErrorCode.COMMENT_NOT_FOUND;
+import static com.prograngers.backend.exception.ErrorCodeBefore.COMMENT_NOT_FOUND;
 
 public class CommentNotFoundException extends NotFoundException {
     public CommentNotFoundException() {
