+++ backend/src/main/java/com/prograngers/backend/exception/notfound/FollowNotFoundException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.notfound;
 
-import static com.prograngers.backend.exception.ErrorCode.NOT_FOUND_FOLLOW;
+import static com.prograngers.backend.exception.ErrorCodeBefore.NOT_FOUND_FOLLOW;
 
 public class FollowNotFoundException extends NotFoundException {
