+++ backend/src/main/java/com/prograngers/backend/exception/notfound/MemberNotFoundException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.notfound;
 
-import static com.prograngers.backend.exception.ErrorCode.MEMBER_NOT_FOUND;
+import static com.prograngers.backend.exception.ErrorCodeBefore.MEMBER_NOT_FOUND;
 
 public class MemberNotFoundException extends NotFoundException {
     public MemberNotFoundException() {
