+++ backend/src/main/java/com/prograngers/backend/exception/notfound/ProblemNotFoundException.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.exception.notfound;
 
 
-import static com.prograngers.backend.exception.ErrorCode.PROBLEM_NOT_FOUND;
+import static com.prograngers.backend.exception.ErrorCodeBefore.PROBLEM_NOT_FOUND;
 
 public class ProblemNotFoundException extends NotFoundException {
