+++ backend/src/main/java/com/prograngers/backend/exception/notfound/ReviewAlreadyDeletedException.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.exception.notfound;
 
 
-import static com.prograngers.backend.exception.ErrorCode.REVIEW_ALREADY_DELETED;
+import static com.prograngers.backend.exception.ErrorCodeBefore.REVIEW_ALREADY_DELETED;
 
 public class ReviewAlreadyDeletedException extends NotFoundException {
     public ReviewAlreadyDeletedException() {
