+++ backend/src/main/java/com/prograngers/backend/exception/notfound/ReviewNotFoundException.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.exception.notfound;
 
 
-import static com.prograngers.backend.exception.ErrorCode.REVIEW_NOT_FOUND;
+import static com.prograngers.backend.exception.ErrorCodeBefore.REVIEW_NOT_FOUND;
 
 public class ReviewNotFoundException extends NotFoundException {
     public ReviewNotFoundException() {
