+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/AlreadyExistMemberException.java
@@ -1,9 +1,9 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import com.prograngers.backend.exception.ErrorCode;
+import com.prograngers.backend.exception.ErrorCodeBefore;
 
 public class AlreadyExistMemberException extends UnAuthorizationException {
     public AlreadyExistMemberException() {
-        super(ErrorCode.ALREADY_EXIST_MEMBER, "이미 존재하는 회원입니다.");
+        super(ErrorCodeBefore.ALREADY_EXIST_MEMBER, "이미 존재하는 회원입니다.");
     }
 }
