+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/AlreadyExistNicknameException.java
@@ -1,9 +1,9 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import com.prograngers.backend.exception.ErrorCode;
+import com.prograngers.backend.exception.ErrorCodeBefore;
 
 public class AlreadyExistNicknameException extends UnAuthorizationException {
     public AlreadyExistNicknameException() {
-        super(ErrorCode.ALREADY_EXIST_NICKNAME, "이미 존재하는 닉네임입니다.");
+        super(ErrorCodeBefore.ALREADY_EXIST_NICKNAME, "이미 존재하는 닉네임입니다.");
     }
 }
