+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/ExpiredTokenException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.EXPIRED_TOKEN;
+import static com.prograngers.backend.exception.ErrorCodeBefore.EXPIRED_TOKEN;
 
 public class ExpiredTokenException extends UnAuthorizationException {
     public ExpiredTokenException() {
