+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/FailToEncodeException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.FAIL_TO_ENCODED;
+import static com.prograngers.backend.exception.ErrorCodeBefore.FAIL_TO_ENCODED;
 
 public class FailToEncodeException extends UnAuthorizationException {
     public FailToEncodeException() {
