+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/FailedSignatureTokenException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.FAILED_SIGNATURE_TOKEN;
+import static com.prograngers.backend.exception.ErrorCodeBefore.FAILED_SIGNATURE_TOKEN;
 
 public class FailedSignatureTokenException extends UnAuthorizationException {
     public FailedSignatureTokenException() {
