+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/IncorrectIssuerTokenException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.NOT_PROGRANGERS_TOKEN;
+import static com.prograngers.backend.exception.ErrorCodeBefore.NOT_PROGRANGERS_TOKEN;
 
 public class IncorrectIssuerTokenException extends UnAuthorizationException {
     public IncorrectIssuerTokenException() {
