+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/IncorrectStateInNaverLoginException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.INCORRECT_NAVER_CODE;
+import static com.prograngers.backend.exception.ErrorCodeBefore.INCORRECT_NAVER_CODE;
 
 public class IncorrectStateInNaverLoginException extends UnAuthorizationException {
     public IncorrectStateInNaverLoginException() {
