+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/InvalidAccessTokenException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.INVALID_ACCESS_TOKEN;
+import static com.prograngers.backend.exception.ErrorCodeBefore.INVALID_ACCESS_TOKEN;
 
 public class InvalidAccessTokenException extends UnAuthorizationException {
     public InvalidAccessTokenException() {
