+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/InvalidClaimTypeException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.INVALID_CLAIM_TYPE;
+import static com.prograngers.backend.exception.ErrorCodeBefore.INVALID_CLAIM_TYPE;
 
 public class InvalidClaimTypeException extends UnAuthorizationException {
     public InvalidClaimTypeException() {
