+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/MemberUnAuthorizedException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.UNAUTHORIZED_MEMBER;
+import static com.prograngers.backend.exception.ErrorCodeBefore.UNAUTHORIZED_MEMBER;
 
 public class MemberUnAuthorizedException extends UnAuthorizationException {
     public MemberUnAuthorizedException() {
