+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/MissingIssuerTokenException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.MISSING_ISSUER_TOKEN;
+import static com.prograngers.backend.exception.ErrorCodeBefore.MISSING_ISSUER_TOKEN;
 
 public class MissingIssuerTokenException extends UnAuthorizationException {
     public MissingIssuerTokenException() {
