+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/NotExistRefreshTokenException.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import static com.prograngers.backend.exception.ErrorCode.NOT_EXIST_REFRESH_TOKEN;
+import static com.prograngers.backend.exception.ErrorCodeBefore.NOT_EXIST_REFRESH_TOKEN;
 
 public class NotExistRefreshTokenException extends UnAuthorizationException {
     public NotExistRefreshTokenException() {
