+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/RefreshTokenNotFoundException.java
@@ -1,9 +1,9 @@
 package com.prograngers.backend.exception.unauthorization;
 
-import com.prograngers.backend.exception.ErrorCode;
+import com.prograngers.backend.exception.ErrorCodeBefore;
 
 public class RefreshTokenNotFoundException extends UnAuthorizationException {
     public RefreshTokenNotFoundException() {
-        super(ErrorCode.NOT_FOUND_REFRESH_TOKEN, "유효한 refresh token이 존재하지 않습니다.");
+        super(ErrorCodeBefore.NOT_FOUND_REFRESH_TOKEN, "유효한 refresh token이 존재하지 않습니다.");
     }
 }
