+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/UnAuthorizationException.java
@@ -1,10 +1,10 @@
 package com.prograngers.backend.exception.unauthorization;
 
 import com.prograngers.backend.exception.CustomException;
-import com.prograngers.backend.exception.ErrorCode;
+import com.prograngers.backend.exception.ErrorCodeBefore;
 
 public class UnAuthorizationException extends CustomException {
-    public UnAuthorizationException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
+    public UnAuthorizationException(ErrorCodeBefore errorCodeBefore, String message) {
+        super(errorCodeBefore, message);
     }
 }
